update design and add postgreSQL
