+++ MnemonicBuilder.Infrastructure/MnemonicBuilder.Infrastructure.csproj
@@ -7,6 +7,13 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="8.0.16" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.16" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="8.0.16" />
+    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.11" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\MnemonicBuilder.Application\MnemonicBuilder.Application.csproj" />
     <ProjectReference Include="..\MnemonicBuilder.Domain\MnemonicBuilder.Domain.csproj" />
   </ItemGroup>
