+++ MnemonicBuilder.Web/Controllers/SentenceController.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using MnemonicBuilder.Application.Words.Queries;
 using MnemonicBuilder.Web.Models;
 
-namespace MnemonicBuilder.Controllers
+namespace MnemonicBuilder.Web.Controllers
 {
     public class SentenceController : Controller
     {
@@ -13,6 +13,8 @@ namespace MnemonicBuilder.Controllers
         {
             _handler = handler;
         }
+
+        
 
         [HttpGet]
         public IActionResult Index()
@@ -25,13 +27,44 @@ namespace MnemonicBuilder.Controllers
         {
             var result = await _handler.Handle(new SearchWordsByPatternQuery(pattern));
 
-            ResultWordsViewModel resultWords = new ResultWordsViewModel
+            SentenceViewModel resultWords = new SentenceViewModel
             {
                 Pattern = pattern,
                 Words = result
             };
 
             return View(resultWords);
+        }
+
+        [HttpGet]
+        public IActionResult Create()
+        {
+            return View();
+        }
+
+        [HttpGet]
+        public IActionResult Edit(int? id)
+        {
+            if (id <= 0 || id == null)
+            {
+                return NotFound();
+            }
+
+            
+
+            return View();
+        }
+
+        [HttpPost]
+        public IActionResult Search(string pattern)
+        {
+             return View();
+        }
+
+        [HttpPost]
+        public IActionResult Save()
+        {
+            return View();
         }
     }
 }
