--- MnemonicBuilder.Web/Models/ResultWordsViewModel.cs
@@ -1,8 +1,0 @@
-﻿namespace MnemonicBuilder.Web.Models
-{
-    public class ResultWordsViewModel
-    {
-        public string Pattern { get; set; }
-        public IEnumerable<string> Words { get; set; }
-    }
-}
