+++ MnemonicBuilder.Web/Views/Sentence/Edit.cshtml
@@ -1,0 +1,7 @@
+﻿
+@{
+    ViewData["Title"] = "Редактировать предложение";
+}
+
+<h1>Edit</h1>
+
