+++ MnemonicBuilder.Web/Views/Sentence/Index.cshtml
@@ -1,71 +1,30 @@
-﻿@model MnemonicBuilder.Web.Models.ResultWordsViewModel
+﻿@model System.Collections.Generic.List<MnemonicBuilder.Web.Models.SentenceViewModel>
 
 @{
-	ViewData["Title"] = "Создать предложение";
+	ViewData["Title"] = "Мои предложения";
 }
 
-<form method="post">
-	<div class="row justify-content-center">
-		<h2 class="text-center">Строитель предложения</h2>
-		<div class="col-md-6 col-lg-5 col-xxl-4">
-			<textarea type="text" name="sentence" value="" class="form-control h-100" rows="5"></textarea>
-		</div>
-		<div class="col-md-6 col-lg-5 col-xxl-4">
-			<div class="row">
-				<div class="align-items-center">
-					<div class="col-md-12 p-1 input-group">
-						<input type="text" asp-for="Pattern" placeholder="Введите регулярное выражение"
-						aria-describedby="search-button" class="form-control" aria-label="Введите регулярное выражение" />
-						<button id="search-button" type="submit" class="btn btn-primary px-3"><i class="bi bi-search"></i></button>
-					</div>
+<div class="row">
+	@foreach (var sentence in new int[] { 1, 2, 3, 4, 5, 6, 7, 8, 9 })
+	{
+		<div class="col-sm-6 col-md-4 col-lg-3 my-2">
+			<div class="card border-info">
+				<div class="card-header">
+					<span class="text-muted">Обновлено: @DateTime.Now.ToShortDateString()</span>
+				</div>
+				<div class="card-body">
+					<h5 class="card-title">Предложение @sentence</h5>
+					<p class="card-text">
+						Каждый Охотник Желает Знать Где Сидит Фазан
+					</p>
+				</div>
+				<div class="card-footer btn-group">
+					<a href="#" class="btn btn-outline-primary"><i class="bi bi-pencil"></i></a>
+					<a href="#" class="btn btn-outline-danger"><i class="bi bi-trash3"></i></a>
 				</div>
 			</div>
-			@* 		<div class="row p-1 text-center">
-			<div class="col-md-6">
-				<button class="btn btn-success px-3"><i class="bi bi-floppy"></i></button>
-			</div>
-			<div class="col-md-6">
-				<button class="btn btn-danger px-3"><i class="bi bi-trash"></i></button>
-			</div>
-		</div> *@
-			<div class="row p-1">
-				<div class="col-12">
-					<label for="words-select">Select list (select one):</label>
-					<select class="form-control" id="sel1">
-						@if (Model != null)
-						{
-							foreach (var word in Model.Words)
-							{
-								<option>@word</option>
-							}
-						}
-						else
-						{
-							<option selected disabled>--Не Найдно--</option>
-						}
-					</select>
-				</div>
 
-			</div>
 		</div>
-		<div class="col-md-12 col-lg-10 col-xxl-8">
-			<h2 class="text-center">Описание предложения</h2>
-			<div class="mb-3">
-				<label for="sentence-name" class="form-label">Название</label>
-				<input type="text" class="form-control" id="sentence-name" required>
-			</div>
-			<div class="mb-3">
-				<label for="sentence" class="form-label">Описание</label>
-				<textarea type="text" name="sentence" value="" class="form-control" rows="4"></textarea>
-			</div>
-			<div class="mb-3 form-check">
-				<input type="checkbox" class="form-check-input" id="privacy-check" checked>
-				<label class="form-check-label" for="privacy-check">Видно остальным пользователям</label>
-			</div>
-			<button type="submit" class="btn btn-primary">Готово</button>
-			<button type="button" class="btn btn-danger">Очистить</button>
-		</div>
-	</div>
-</form>
+	}
 
-
+</div>
