+++ MnemonicBuilder.Web/Views/Shared/_Layout.cshtml
@@ -21,7 +21,7 @@
                 <div class="navbar-collapse collapse d-sm-inline-flex justify-content-between">
                     <ul class="navbar-nav flex-grow-1">
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="home" asp-action="index">Мои предложения</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="sentence" asp-action="index">Мои предложения</a>
                         </li>
                     </ul>
                     <partial name="_LoginPartial" />
