+++ MnemonicBuilder.Web/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
-﻿@using MnemonicBuilder
-@using MnemonicBuilder.Models
+﻿@using MnemonicBuilder.Web
+@using MnemonicBuilder.Web.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
