+++ MnemonicBuilder.Web/appsettings.json
@@ -1,7 +1,7 @@
 {
-  "ConnectionStrings": {
-    "DefaultConnection": "Server=(localdb)\\mssqllocaldb;Database=aspnet-MnemonicBuilder-a5b69eb5-3a4d-457a-992f-bbf83decccf5;Trusted_Connection=True;MultipleActiveResultSets=true"
-  },
+    "ConnectionStrings": {
+        "DefaultConnection": "Host=localhost;Database=Mnemonic;Username=postgres;Password=4650390"
+    },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
