+++ MnemonicBuilder.Web/wwwroot/data/all_russian_words.txt
@@ -1,18 +1,15 @@
-арбуз
-азбука
-банан
-чебурек
-охотник
-апельсин
-баран
-крендель
-ручка
-телефон
-фазан
-барс
-карандаш
-коробка
-ножницы
-hello
-world
-hey
+Р°СЂР±СѓР·
+Р°Р·Р±СѓРєР°
+Р±Р°РЅР°РЅ
+С‡РµР±СѓСЂРµРє
+РѕС…РѕС‚РЅРёРє
+Р°РїРµР»СЊСЃРёРЅ
+Р±Р°СЂР°РЅ
+РєСЂРµРЅРґРµР»СЊ
+СЂСѓС‡РєР°
+С‚РµР»РµС„РѕРЅ
+С„Р°Р·Р°РЅ
+Р±Р°СЂСЃ
+РєР°СЂР°РЅРґР°С€
+РєРѕСЂРѕР±РєР°
+РЅРѕР¶РЅРёС†С‹
